risoluzione jira-105 duplicazione finestra di caricamento dei layer di un servizio
