+++ frontend/localConfig.json
@@ -160,7 +160,7 @@
 		   {
 			   "name": "Search",
 			   "cfg": {
-				 "withToggle": ["max-width: 768px", "min-width: 768px"]
+				 "withToggle": true
 			   }
 		   },
             {
